[kina.lee] 로또 4단계 구현 (#414)
* 프로젝트 생성 및 초기 설정

* gradle dependency 추가(assertj)

* Application 클래스 추가

* 테스트 클래스 추가

* Prize 열거체 추가

* Lotto 클래스 추가

* LottoResult 클래스 추가

* CustomScanner 클래스 추가

* LottoCliView 클래스 추가

* LottoManager 클래스 추가

* 코드 정리

* 수익률 추가

* Prize 열거체를 Prize, Rank로 세분화

* CustomScanner에서 출력문 제거

* 입력값 검증, Lotto 결과 확인을 위한 LottoChecker 클래스 추가

* LottoResult 클래스 재구성

로또 결과와 관계없는 멤버 제거

* LottoMessage 열거체 추가

프로그램에서 사용되는 문자열을 LottoMessage에서 관리하도록 변경.

* LottoManager 클래스 변경

* 수동 구매시 출력되는 메시지 추가

* 수동 구매시 사용되는 생성자 추가

* 수동 구매 기능 추가

* Lotto를 상속받는 WinningLotto 클래스 추가

* 코드 리뷰 내용 반영

* sparkjava 사용을 위한 의존성 추가

* 코드 정리

* 수익률 계산 식 변경

* 웹 UI 적용

+++ src/main/java/lotto/domain/constant/Prize.java
@@ -11,9 +11,9 @@ public enum Prize {
     FAILED(Rank.FAILED, 0);
 
     private final Rank rank;
-    private final int winnings;
+    private final long winnings;
 
-    Prize(Rank rank, int winnings) {
+    Prize(Rank rank, long winnings) {
         this.rank = rank;
         this.winnings = winnings;
     }
@@ -22,7 +22,7 @@ public enum Prize {
         return rank;
     }
 
-    public int getWinnings() {
+    public long getWinnings() {
         return winnings;
     }
